ADD testIds and test dapp pom
