+++ packages/e2e-tests/loginWithDapp/models/TestDapp.js
@@ -1,0 +1,31 @@
+const { testDappURL } = require("../../utils/config");
+
+class TestDappPage {
+    constructor(page) {
+        this.page = page;
+    }
+    async navigate() {
+        await this.page.goto(testDappURL);
+    }
+    async clickSignIn() {
+        await this.page.click("data-test-id=testDapp-signInBtn");
+    }
+    async getAccessKeyForAccountId(accountId) {
+        return this.page.evaluate(
+            (accountId) =>
+                Object.keys(window.localStorage).find((k) =>
+                    new RegExp(accountId).test(k)
+                ),
+            accountId
+        );
+    }
+    async getPendingAccessKeys() {
+        return this.page.evaluate(() =>
+            Object.keys(window.localStorage).filter((k) =>
+                /pending_key/.test(k)
+            )
+        );
+    }
+}
+
+module.exports = { TestDappPage };
